Add XZ decoder.
This is probably not optimally efficient, but seems fast enough decoding some real world .xz files.
The implementation was done mostly by looking at the Linux kernel implementation (public domain) and
restructuring it to be easier to understand.

Not that BCJ filters are not implemented, but could be added later if necessary.

Fixes https://github.com/brendan-duncan/archive/issues/94